feat: use chevrotain
