+++ src/bsl/codeModel/index.ts
@@ -1,4 +1,3 @@
 export * from './model'
-export * from './bslCodeModelProvider'
 export * from './interfaces'
 export * from './visitor'
