+++ src/bsl/editor/editor.ts
@@ -1,5 +1,7 @@
 import { editor } from 'monaco-editor-core'
 import { EditorScope } from '@/bsl/scope/editorScope'
+import { ChevrotainModuleModel } from '../chevrotain/moduleModel'
+import { TreeSitterModuleModel } from '../tree-sitter/moduleModel'
 
 let activeEditor: BslEditor | undefined
 
@@ -17,7 +19,6 @@ export class BslEditor {
         if (container === null) {
             throw 'Error!';
         }
-
         this.editor = editor.create(container, {
             language: 'bsl',
             automaticLayout: true,
@@ -38,9 +39,10 @@ export class BslEditor {
             // fontFamily: 'Courier New',
             // fontSize: 12,
             fontLigatures: true,
-            wordBasedSuggestions: false
+            wordBasedSuggestions: false,
+            model: this.createModel()
         });
-        
+
         tuneEditor(this.editor)
 
         this.scope = EditorScope.createScope(this.editor)
@@ -50,17 +52,28 @@ export class BslEditor {
         })
     }
 
-    set content(value:string){
+    set content(value: string) {
         this.editor.setValue(value)
     }
 
-    get content(){
+    get content() {
         return this.editor.getValue()
     }
 
     getText(): string {
         const model = this.editor.getModel()
         return model ? model.getValue() : ''
+    }
+
+    createModel() {
+
+        const model = editor.createModel('', 'bsl');
+
+        if (wasmSupported()) {
+            return TreeSitterModuleModel.create(model)
+        } else {
+            return ChevrotainModuleModel.create(model)
+        }
     }
 }
 
@@ -73,4 +86,9 @@ function tuneEditor(editor: editor.IStandaloneCodeEditor) {
 
 export function getActiveEditor() {
     return activeEditor
+}
+
+function wasmSupported() {
+    return false && typeof WebAssembly === "object"
+        && typeof WebAssembly.instantiate === "function"
 }
