+++ src/bsl/serviceRegistry.ts
@@ -1,7 +1,12 @@
-import { BslCodeModelProvider, provider } from "./codeModel/bslCodeModelProvider"
+import { BslCodeModelProvider, provider, factory } from "./codeModel/bslCodeModelProvider"
+import { CodeModelFactory } from "./codeModel/codeModelFactory"
 
 export default {
     get codeModelProvider(): BslCodeModelProvider {
         return provider
+    },
+
+    get codeModelFactory(): CodeModelFactory | undefined {
+        return factory
     }
 }
