+++ src/bsl/tree-sitter/moduleModel.ts
@@ -1,79 +1,50 @@
-import { LocalModuleScope, Member, MemberType, MethodScope } from "@/common/scope"
 import { BslParser } from "./bslParser"
 import { editor, Position } from "monaco-editor-core"
-import { Method, Variable } from "@/common/codeModel"
 import { ModuleModel } from "../moduleModel"
 import { ExpressionProvider } from "../expressions/expressionProvider"
 import { Constructor, Expression, MethodCall } from "../expressions/expressions"
-import { resolveMethodSymbol, resolveSymbol } from "./symbols"
 import { getEditedPositionOffset, getPositionOffset } from "@/monaco/utils"
+import { resolveMethodSymbol, resolveSymbol } from "./expressionsFactory"
+import { TreeSitterCodeModelFactory } from "."
+import { AutoDisposable } from "@/common/utils/autodisposable"
+import { BslCodeModel } from "../codeModel"
+import { BslModuleScope } from "../scope/bslModuleScope"
 
-export class TreeSitterModuleModel extends LocalModuleScope implements ExpressionProvider {
+export class TreeSitterModuleModel extends AutoDisposable implements ExpressionProvider {
 
     static create(editorModel: editor.ITextModel): ModuleModel {
         const moduleScope = new TreeSitterModuleModel(editorModel);
 
-        (editorModel as ModuleModel).getScope = () => moduleScope;
+        (editorModel as ModuleModel).getScope = moduleScope.getScope.bind(moduleScope);
         (editorModel as ModuleModel).getEditingExpression = moduleScope.getEditingExpression.bind(moduleScope);
         (editorModel as ModuleModel).getCurrentExpression = moduleScope.getCurrentExpression.bind(moduleScope);
         (editorModel as ModuleModel).getEditingMethod = moduleScope.getEditingMethod.bind(moduleScope);
+        (editorModel as ModuleModel).getCodeModel = () => moduleScope.codeModel;
 
         return editorModel as ModuleModel
     }
 
     parser: BslParser
+    codeModel: BslCodeModel | undefined
+    editorModel: ModuleModel
+    scope: BslModuleScope
 
     constructor(model: editor.ITextModel) {
-        super(model)
+        super()
+        this.editorModel = model as ModuleModel
+        this.scope = new BslModuleScope(this.editorModel)
+        
         this._disposables.push(this.parser = new BslParser(model))
+        
+        this.codeModel = TreeSitterCodeModelFactory.buildModel(this.parser)
     }
 
-    didUpdateMembers() {
-        this.module.methods = this.parser.methods()
-        this.module.vars = this.parser.vars()
-        this.members.length = 0
-
-        for (let i = 0; i < this.module.methods.length; i++) {
-            this.members.push({
-                kind: MemberType.function,
-                name: this.module.methods[i].name
-            })
-        }
-        for (let i = 0; i < this.module.vars.length; i++) {
-            this.members.push({
-                kind: MemberType.property,
-                name: this.module.vars[i].name
-            })
-        }
-    }
-
-    protected createMethodScope(method: Method): MethodScope {
-        const members: Member[] = []
-
-        if (!method.vars) {
-            method.vars = []
-            const iter = this.parser.getMethodVars(method)
-            let variable: IteratorResult<Variable>
-            while (!(variable = iter.next()).done) {
-                method.vars.push(variable.value)
-            }
-        }
-
-        method.vars.forEach(v => members.push({
-            name: v.name,
-            kind: MemberType.variable,
-            type: v.type
-        }))
-        method.params.forEach(v => members.push({
-            name: v.name,
-            kind: MemberType.variable,
-        }))
-
-        return new MethodScope(members)
+    getScope() {
+        return this.scope
     }
 
     getEditingExpression(position: Position): Expression | undefined {
-        const positionOffset = getEditedPositionOffset(this.model, position)
+        const positionOffset = getEditedPositionOffset(this.editorModel, position)
 
         const node = this.parser.getCurrentEditingNode(positionOffset)
         if (node) {
@@ -82,14 +53,14 @@ export class TreeSitterModuleModel extends LocalModuleScope implements Expressio
     }
 
     getCurrentExpression(position: Position): Expression | undefined {
-        const node = this.parser.getCurrentNode(getPositionOffset(this.model, position))
+        const node = this.parser.getCurrentNode(getPositionOffset(this.editorModel, position))
         if (node) {
             return resolveSymbol(node)
         }
     }
 
     getEditingMethod(position: Position): Constructor | MethodCall | undefined {
-        const positionOffset = getEditedPositionOffset(this.model, position)
+        const positionOffset = getEditedPositionOffset(this.editorModel, position)
         const node = this.parser.getCurrentEditingNode(positionOffset)
         if (node) {
             return resolveMethodSymbol(node)
