--- src/bsl/tree-sitter/symbols/index.ts
@@ -1,1 +1,0 @@
-export * from './factory'
