+++ src/common/scope/index.ts
@@ -2,5 +2,4 @@ export * from './model/members'
 export * from './model/scope'
 export * from './model/types'
 export * from './globalScopeManager'
-export * from './localModuleScope'
 export * from './methodScope'
